Spravený "konštuktor", "model", "napájací_palivový systém", "simulácia". Simulujeme takmer všetky porebné parametre skonštruovanej lodi.
